# docs/swagger.yaml
basePath: /
definitions:
...
  event.FormConfigs:
    properties:
      email:
        $ref: '#/definitions/event.FormEmailConfig'
      redirect:
        $ref: '#/definitions/event.FormRedirectConfig'
      refundPolicy:
        $ref: '#/definitions/event.FormRefundPolicyConfig'
    type: object
  event.FormEmailConfig:
    properties:
      content:
        type: string
      enabled:
        type: boolean
    type: object
  event.FormLink:
    properties:
      description:
        type: string
      label:
        type: string
      message:
        type: string
      phone:
        type: string
      url:
        type: string
    type: object
  event.FormRedirectConfig:
    properties:
      enabled:
        type: boolean
      link:
        $ref: '#/definitions/event.FormLink'
      type:
        type: string
    type: object
  event.FormRefundPolicyConfig:
    properties:
      daysBeforeEvent:
        type: integer
      enabled:
        type: boolean
      refundPercentage:
        type: integer
      refundType:
        type: string
    type: object
  event.StatusRequest:
    properties:
      status:
        type: string
    type: object
  event.UpdateEventImageRequest:
    properties:
      eventId:
        type: string
      imageUrl:
        type: string
    type: object
  event.UpdateEventVideoRequest:
    properties:
      eventId:
        type: string
      videoUrl:
        type: string
    type: object
...
  github_com_gbaski_gbaski-host_internal_event.CreateEventRequest:
    properties:
      accessType:
        type: string
      categoryId:
        type: integer
      description:
        type: string
      durationType:
        type: string
      endDate:
        type: string
      formId:
        type: string
      imageUrl:
        type: string
      location:
        type: string
      modeType:
        type: string
      name:
        type: string
      otherCategory:
        type: string
      payment:
        type: string
      regType:
        type: string
      socialMedia:
        items:
          type: integer
        type: array
      startDate:
        type: string
      videoUrl:
        type: string
    required:
    - description
    - durationType
    - endDate
    - modeType
    - name
    - payment
    - regType
    - startDate
    type: object
  github_com_gbaski_gbaski-host_internal_event.FormRequest:
    properties:
      accessType:
        type: string
      configs:
        $ref: '#/definitions/event.FormConfigs'
      currency:
        type: string
      endDate:
        type: string
      eventId:
        type: string
      feePayer:
        type: string
      id:
        type: string
      isPrimaryForm:
        type: boolean
      name:
        type: string
      price:
        type: number
      regType:
        type: string
      schema:
        items:
          type: integer
        type: array
      startDate:
        type: string
    type: object
  github_com_gbaski_gbaski-host_internal_event.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      otp_required:
        type: boolean
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gbaski.app
    name: API Support
  description: API documentation for Gbaski Host service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gbaski Host API
  version: "1.0"
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of events for the authenticated user
        with optional filtering and search
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter parameters
        in: query
        name: filter
        type: string
      - description: Range parameters
        in: query
        name: range
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get all events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event for the authenticated user
      parameters:
      - description: Event creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Create new event
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event by ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Delete event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Retrieve a single event by its ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update an existing event by ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Event update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event
      tags:
      - events
  /events/{id}/stats:
    get:
      consumes:
      - application/json
      description: Retrieve statistics for a specific event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event statistics
      tags:
      - events
  /events/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an existing event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event status
      tags:
      - events
  /events/categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available event categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event categories
      tags:
      - events
  /events/image:
    post:
      consumes:
      - application/json
      description: Update the image URL for an event
      parameters:
      - description: Event image update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.UpdateEventImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event image
      tags:
      - events
  /events/report:
    get:
      consumes:
      - application/json
      description: Retrieve a report of all events for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get events report
      tags:
      - events
  /events/video:
    post:
      consumes:
      - application/json
      description: Update the video URL for an event
      parameters:
      - description: Event video update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.UpdateEventVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event video
      tags:
      - events
...
  /registrations/forms:
    get:
      consumes:
      - application/json
      description: Retrieve all registration forms for the authenticated user's events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event forms
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new registration form for an event
      parameters:
      - description: Form creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.FormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Create registration form
      tags:
      - events
...
schemes:
- http
- https
swagger: "2.0"